+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { Provider } from "react-redux";
-import Context from "./containers/Context/Context";
+import Content from "./containers/Context/Content";
 import store from "./redux/store";
 import "react-toastify/dist/ReactToastify.css";
 import { ToastContainer } from "react-toastify";
@@ -7,7 +7,7 @@ import { ToastContainer } from "react-toastify";
 const App = () => {
   return (
     <Provider store={store}>
-      <Context />
+      <Content />
       <ToastContainer />
     </Provider>
   );
