+++ src/components/DiaryCard/DiaryCard.tsx
@@ -1,7 +1,12 @@
 import { useState } from "react";
 import { Box, Button, Grid, Typography } from "@mui/material";
-import { DiaryCardProps } from "../../types/DiaryCardProps";
 import { normalCardColor, specialCardColor } from "../../theme/colors";
+
+export type DiaryCardProps = {
+  title: string;
+  name: string;
+  description: string;
+};
 
 const DiaryCard = (props: DiaryCardProps) => {
   const [isView, setIsView] = useState<boolean>(false);
@@ -34,7 +39,7 @@ const DiaryCard = (props: DiaryCardProps) => {
             ? props.description.slice(0, 5) + "..."
             : props.description}
         </Typography>
-        {props.description.length > 5 ? (
+        {props.description.length > 5 && (
           <Grid container justifyContent="flex-end">
             <Button
               sx={{ justifyContent: "end" }}
@@ -52,8 +57,6 @@ const DiaryCard = (props: DiaryCardProps) => {
               )}
             </Button>
           </Grid>
-        ) : (
-          <></>
         )}
       </Box>
     </Grid>
