--- src/types/DiaryCardProps.ts
@@ -1,5 +1,0 @@
-export type DiaryCardProps = {
-  title: string;
-  name: string;
-  description: string;
-};
