--- src/types/DiaryData.ts
@@ -1,5 +1,0 @@
-export type DiaryData = {
-  title: string;
-  username: string;
-  description: string;
-};
