+++ src/utils/RoutesUtil.ts
@@ -1,0 +1,2 @@
+export const homeRoute = "/home";
+export const defaultRoute = "/";
